==> 1_scrape_eurovision.ipynb <==
# Scraping the Eurovision Song Contest data from 1956 until 2019

import numpy as np
import pandas as pd
import requests
from bs4 import BeautifulSoup
from pprint import pprint
import re
%matplotlib inline

### DataFrame of Eurovision Song Contest Winners per Year (We did not use this list)

url = 'https://en.wikipedia.org/wiki/List_of_Eurovision_Song_Contest_winners'
res = requests.get(url)
#res
soup = BeautifulSoup(res.content, 'html.parser')
#soup
table = soup.body.find_all('table', {'class':'wikitable sortable'})[0]
#table
rows = table.find_all('tr')
#rows
rows = [row.text.strip().split('\n\n') for row in rows]
#rows
colnames = rows[0]
data = rows[1:]
df = pd.DataFrame(data, columns=colnames)
df.head(10)
df.to_csv('winners_per_year.csv', index=False)
check = pd.read_csv('winners_per_year.csv')
check.head()

### DataFrame of Eurovision Song Contest per Year (I cannot scrape this table)

year = 2019
url = f'https://eurovisionworld.com/eurovision/{year}'
res = requests.get(url)
res
soup = BeautifulSoup(res.content, 'html.parser')
soup

#try1
table = soup.body.main
table

#try2
table = soup.body.main.find('div', {'id':'voting_table'})
table

#try3
table = soup.body.main.find_all('table', {'class':'voting_table'})
table

#CANNOT SCRAPE THE TABLE

### Eurovision Song Contest Website (Did not work for all the different years)

#try1
url_end = ['tel-aviv-2019/grand-final', 'lisbon-2018/grand-final', 'kyiv-2017/grand-final', '', 'hilversum-1958/final', 'frankfurt-1957/final', 'lugano-1956/final']
url = f'https://eurovision.tv/event/{url_end[2]}'
res = requests.get(url)
#res
soup = BeautifulSoup(res.content, 'html.parser')
#soup
table = soup.body.main.find_all('table', {'class': 'event-table'})[0]
#table
rows = table.find_all('tr')
rows = [row.text.strip().split('\n') for row in rows]
rows[1:]
data = rows[1:]
df = pd.DataFrame(data)
pd.options.display.max_columns = None
df.head()
df = df[[0,7,15,21,27,30]].head(10)
df = df.rename(columns={0:'R/O', 7:'Country', 15:'Contestant', 21:'Song', 27:'Points', 30:'Place'})
df.head()
# DOES NOT WORK WITH ALL THE URLS

## Eurovision Song Contest Website (Try2)

#There are different url ends for the different years.
#There is a pattern, however, the end is '/grand-final' from 2004, before 2004 it is '\final' 
url_end = ['tel-aviv-2019/grand-final', 'lisbon-2018/grand-final', 'kyiv-2017/grand-final', 'hilversum-1958/final', 'frankfurt-1957/final', 'lugano-1956/final']
url = f'https://eurovision.tv/event/{url_end[0]}'
res = requests.get(url)
soup = BeautifulSoup(res.content, 'html.parser')
table = soup.body.main.find_all('table', {'class': 'event-table'})[0]
rows = table.find_all('tr')
#rows = [row.text.strip().split('\n') for row in rows]
sample_row = rows[2]
sample_row

### Functions for scraping data of the columns of the Scorecard table

# get R/O
#sample_row.find('td').text.strip()

def get_ro(row):
    return row.find('td').text.strip()
    
get_ro(sample_row)

# get country
#sample_row.span.text

def get_country(row):
    return row.span.text

get_country(sample_row)

# get contestant
#sample_row.find_all('a')[1].text.strip()

def get_contestant(row):
    return row.find_all('a')[1].text.strip()

get_contestant(sample_row)

# get song
#sample_row.find_all('div')[3].text.strip()

def get_song(row):
    return row.find_all('div')[3].text.strip()

get_song(sample_row)

# get points
#sample_row.find_all('td', class_='text-right')[0].text.strip()

def get_points(row):
    return row.find_all('td', class_='text-right')[0].text.strip()

get_points(sample_row)

# get place
#sample_row.find_all('td', class_='text-right')[1].text.strip()

def get_place(row):
    return row.find_all('td', class_='text-right')[1].text.strip()
    
get_place(sample_row)

### Functions for scraping the year and the host city from the urls

# get year
#url

def get_year(url):
    year_pattern = r'(\d{4})'
    return int(re.findall(year_pattern, url)[0])

get_year(url)

#get host city
#url

def get_host_city(url):
    #host_city_pattern = r'(\w+)-'
    host_city_pattern = r'([a-z-]*[a-z]+)-'
    host_city = re.findall(host_city_pattern, url)[0]
    host_city = host_city.title()
    return host_city
    
get_host_city(url)

### Put the scraping functions together and create a dictionary

def create_row(row, url):
    return {'R/O': get_ro(row),
            'Year': get_year(url),
            'Host_City': get_host_city(url),
            'Country': get_country(row),
            'Contestant': get_contestant(row),
            'Song': get_song(row),
            'Points': get_points(row),
            'Place': get_place(row)}

create_row(sample_row, url)

### Create a sample DataFrame

data = [create_row(row, url) for row in rows[1:]] # don't want to get the first row, that is the header

df_from_data = pd.DataFrame(data)
df_from_data.head()

## Put everything together

### 1. Get all the URLs from 1956 until 2019

def get_eurovision_urls():
    urls_source = 'https://eurovision.tv/events'
    res = requests.get(urls_source)
    soup = BeautifulSoup(res.content, 'html.parser')
    a_tags1 = soup.find_all('a', class_='text-blue-darkest')[2:18]
    urls1 = [url['href']+'/grand-final' for url in a_tags1]
    a_tags2 = soup.find_all('a', class_='text-blue-darkest')[18:]
    urls2 = [url['href']+'/final' for url in a_tags2]
    urls = urls1 + urls2
    return urls

urls = get_eurovision_urls()
urls[0]

### 2. A funtion that creates a DataFrame after scraping the URLs

def scrape_eurovision(url):
    res = requests.get(url)
    soup = BeautifulSoup(res.content, 'html.parser')
    table = soup.body.main.find_all('table', {'class': 'event-table'})[0]
    rows = table.find_all('tr')
    data = [create_row(row, url) for row in rows[1:]] # don't want to get the first row, that is the header
    df_from_data = pd.DataFrame(data)
    return df_from_data

# test the function
frame1 = scrape_eurovision(urls[0])

### 3. Run the function for all the URLs

frames = [scrape_eurovision(url) for url in urls]

### 4. Concat all the DataFrames of the different years

result = pd.concat(frames, ignore_index=True)
result.head()

result.tail()

### 5. Save the final result as a CSV file

result.to_csv('eurovision_per_year.csv', index=False)
check = pd.read_csv('eurovision_per_year.csv')
check.head(10)

check.tail(10)
