==> 2_Spotify_song_search.ipynb <==
import spotipy
import song_getter as sg
import pandas as pd
import json
import os

#Creating a connection to the API
#A Spotify account is required for this in order to obtain a client_id and client_secret
#These can be obtained by accessing https://developer.spotify.com/dashboard

#They should be stored as environment variables SPOTIFY_CLIENT_ID and SPOTIFY_CLIENT_SECRET 
#prior to running the code in this notebook

client_id = os.getenv("SPOTIFY_CLIENT_ID")
client_secret = os.getenv("SPOTIFY_CLIENT_SECRET")

from spotipy.oauth2 import SpotifyClientCredentials

client_credentials_manager = SpotifyClientCredentials(client_id=client_id,
                                                      client_secret=client_secret)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

#Reading the scraped Eurovision song data 

eurovision = pd.read_csv("data/eurovision_per_year.csv")

eurovision.head()

#Retrieving the track information for each song from the Spotify API, including track URI

results = eurovision["Song"].apply(lambda x: sg.df_from_search(sp, x))

results.head()

results[0]

#Saving a dataframe with several nested dataframes to a csv was too messy, so we chose to save the search results
#as a separate file. The serach results were saved as a single concatenated dataframe.

#The ID in the ID column matches the row index in the eurivision dataframe and allows dividing the concatenated 
#dataframe into the search results for individual songs.

#A row with NAs was added where the search yielded no results, in order to keep the IDs consistent.

empty_df = pd.DataFrame()
empty_df["track_name"] = [pd.NA]
empty_df["artist_name"] = [pd.NA]
empty_df["uri"] = [pd.NA]
empty_df

list_of_dfs = []
for i, j in enumerate(results):
    if pd.notna(search_results_with_uri["uri"][i]):
        list_of_dfs.append(j.assign(id=i))
    else:
        list_of_dfs.append(empty_df)

search_results_concat = pd.concat(list_of_dfs)

search_results_concat.to_csv("data/search_results_concat.csv")

search_results_concat.head(25)

#Adding the results list as a new column 

query_search_results = eurovision.assign(query_search_results = results)

#Creating a new series with the right track URI for each song

search_results_with_uri = query_search_results.apply(lambda x: sg.find_right_track(x["query_search_results"] , x["Contestant"]), axis=1)

#Adding that series as a new column to the dataframe

search_results_with_uri = query_search_results.assign(uri = search_results_with_uri)

#Dropping the column with search results and saving it to a csv file

eurovision_with_uris = search_results_with_uri.drop("query_search_results", axis="columns")

eurovision_with_uris.to_csv("data/eurovision_with_uris.csv")

eurovision_with_uris.head()

print(f"We retrieved URIs for {sum(eurovision_with_uris['uri'].notna())} songs out of a total of \
{len(eurovision_with_uris['uri'])}")
