==> 3_Spotify_audio_features.ipynb <==
# libraries

import pandas as pd
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

# environment variables

client_id = "3bb2acc187d54269bd4d8a8e29abcb42"
client_secret = "76bf182d940c4b52b51272f863f56692"

# credential manager 

client_credentials_manager = SpotifyClientCredentials(client_id = client_id, client_secret = client_secret)
spotify = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

# create data frame with year/song/uri

df = pd.read_csv("eurovision_with_uris.csv")
df.head()

df.columns #type(tracks['uri'])

tracks = list(df['uri']) #convert series into a list
tracks[0:10]

# run through the list of uris and create list of features for ech uri

client_credentials_manager = SpotifyClientCredentials(client_id = client_id, client_secret = client_secret)
spotify = spotipy.Spotify(client_credentials_manager=client_credentials_manager)


def get_features(tracks):
    list_features =[]
    for track in tracks:
        if type(track)==str:
            feature = spotify.audio_features(track) #search query
            list_features.append(feature)
        else:
            list_features.append(pd.NA)
    return list_features

list_features = get_features(tracks)
list_features[0:20]

# convert list of features into a dataframe

list_of_series = []
for i in list_features:
    empty_series = pd.Series({  "danceability" : pd.NA,
                                "energy" : pd.NA,
                                "key" : pd.NA,
                                "loudness" : pd.NA,
                                "mode" : pd.NA,
                                "speechiness" : pd.NA,
                                "acousticness" : pd.NA,
                                "instrumentalness" : pd.NA,
                                "liveness" : pd.NA,
                                "valence" : pd.NA,
                                "tempo" : pd.NA,
                                "type" : pd.NA,
                                "id" : pd.NA,
                                "uri" : pd.NA,
                                "track_href" : pd.NA,
                                "analysis_url" : pd.NA,
                                "duration_ms" : pd.NA,
                                "time_signature" : pd.NA})
    if pd.notna(i):
        list_of_series.append(pd.Series(i[0]))
    else:
        list_of_series.append(empty_series)

features_df = pd.DataFrame(list_of_series[0]).transpose()

for i in range(1, len(list_of_series)):
    features_df = features_df.append(list_of_series[i], ignore_index=True)
    
features_df.head(20)

# create csv file

features_df.to_csv("data/list_features.csv")
